moved parser to utils.js
